Documentation: share --jobs option ???
The fetch, pull, and clone builtins all have some form of a `--jobs`
command-line option.

The fetch and pull builtins both share the same text (via an include of
fetch-options.txt, where the option's documentation lives), but there is
some repetition in clone's documentation.

Remove this duplication by sharing the `--jobs` option among fetch,
pull, and clone by moving it to the new "fetch-options-shared.txt".  A
couple of small gotchas:

  - The push builtin does not have a --jobs option, so --jobs is
    correctly omitted from git-push(1), despite including the shared
    fetch options documentation.

  - The language between fetch and pull's documentation for the --jobs
    option varies slightly. The first couple of sentences are roughly
    the same (the one in fetch-options.txt omits the default behavior),
    and this is reflected in the patch below.

    But the documentation for the clone builtin does not include the
    following paragraphs (pertaining to the `--multiple` option, which
    does not exist for the clone builtin). These too are reflected in
    the patch below.

Signed-off-by: Taylor Blau <me@ttaylorr.com>